+++ internal/plugins/collector/dns/dns.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/hidracloud/hidra/v3/internal/metrics"
 	"github.com/hidracloud/hidra/v3/internal/misc"
+	"github.com/hidracloud/hidra/v3/internal/plugins"
 	"github.com/hidracloud/hidra/v3/internal/utils"
-	"github.com/hidracloud/hidra/v3/plugins"
 
 	"github.com/likexian/whois"
 	whoisparser "github.com/likexian/whois-parser"
