+++ internal/plugins/collector/ftp/ftp.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/hidracloud/hidra/v3/internal/metrics"
 	"github.com/hidracloud/hidra/v3/internal/misc"
-	"github.com/hidracloud/hidra/v3/plugins"
+	"github.com/hidracloud/hidra/v3/internal/plugins"
 	ftpclient "github.com/jlaffaye/ftp"
 )
