+++ internal/plugins/collector/udp/udp_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"testing"
 
-	"github.com/hidracloud/hidra/v3/plugins"
-	"github.com/hidracloud/hidra/v3/plugins/collector/udp"
+	"github.com/hidracloud/hidra/v3/internal/plugins"
+	"github.com/hidracloud/hidra/v3/internal/plugins/collector/udp"
 )
 
 func TestScenario(t *testing.T) {
