+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/hidracloud/hidra/v3/cmd"
-	_ "github.com/hidracloud/hidra/v3/plugins/all"
+	_ "github.com/hidracloud/hidra/v3/internal/plugins/all"
 )
 
 func main() {
